requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml --output-file=requirements/base.txt requirements/base.in
#
aiohttp==3.8.5
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
anyio==3.7.1
    # via httpcore
argon2-cffi==21.3.0
    # via -r requirements/base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.7.2
    # via
    #   django
    #   django-htmx
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
billiard==4.1.0
    # via celery
celery==5.3.4
    # via
    #   -r requirements/base.in
    #   django-celery-beat
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
    #   gevent
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
crispy-tailwind==1.0.3
    # via -r requirements/base.in
cron-descriptor==1.4.0
    # via django-celery-beat
cryptography==41.0.2
    # via
    #   jwcrypto
    #   pyjwt
decorator==5.1.1
    # via jsonpath-ng
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.14
    # via jwcrypto
django==4.2.5
    # via
    #   -r requirements/base.in
    #   crispy-tailwind
    #   django-allauth
    #   django-autocomplete-light
    #   django-celery-beat
    #   django-cors-headers
    #   django-crispy-forms
    #   django-filter
    #   django-htmx
    #   django-model-utils
    #   django-oauth-toolkit
    #   django-redis
    #   django-tables2
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.54.0
    # via -r requirements/base.in
django-autocomplete-light==3.11.0
    # via -r requirements/base.in
django-celery-beat==2.5.0
    # via -r requirements/base.in
django-cors-headers==4.2.0
    # via -r requirements/base.in
django-crispy-forms==2.0
    # via
    #   -r requirements/base.in
    #   crispy-tailwind
django-environ==0.10.0
    # via -r requirements/base.in
django-filter==24.3
    # via -r requirements/base.in
django-htmx==1.19.0
    # via -r requirements/base.in
django-model-utils==4.3.1
    # via -r requirements/base.in
django-oauth-toolkit==2.3.0
    # via -r requirements/base.in
django-redis==5.3.0
    # via -r requirements/base.in
django-tables2==2.6.0
    # via -r requirements/base.in
django-timezone-field==5.1
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
drf-spectacular==0.26.4
    # via -r requirements/base.in
et-xmlfile==1.1.0
    # via openpyxl
flatten-dict==0.4.2
    # via -r requirements/base.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
geographiclib==2.0
    # via geopy
geopy==2.4.1
    # via -r requirements/base.in
gevent==24.11.1
    # via -r requirements/base.in
greenlet==3.1.1
    # via gevent
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hiredis==2.2.3
    # via -r requirements/base.in
hpack==4.0.0
    # via h2
httpcore==0.17.3
    # via httpx
httpx[http2]==0.24.1
    # via -r requirements/base.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
jsonpath-ng==1.5.3
    # via -r requirements/base.in
jsonschema==4.18.4
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
jwcrypto==1.5.0
    # via django-oauth-toolkit
kombu==5.3.2
    # via celery
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==2.3.4
    # via pandas
oauthlib==3.2.2
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
openpyxl==3.1.2
    # via tablib
pandas==2.3.3
    # via -r requirements/base.in
pillow==10.0.0
    # via -r requirements/base.in
ply==3.11
    # via jsonpath-ng
prompt-toolkit==3.0.39
    # via click-repl
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   django-allauth
    #   twilio
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   celery
    #   pandas
    #   python-crontab
python-slugify==8.0.1
    # via -r requirements/base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   django-timezone-field
    #   djangorestframework
    #   pandas
    #   twilio
pyyaml==6.0.1
    # via drf-spectacular
quickcache==0.5.4
    # via -r requirements/base.in
redis==4.6.0
    # via
    #   -r requirements/base.in
    #   django-redis
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-allauth
    #   django-oauth-toolkit
    #   requests-oauthlib
    #   twilio
requests-oauthlib==1.3.1
    # via django-allauth
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   flatten-dict
    #   jsonpath-ng
    #   python-dateutil
    #   quickcache
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.4
    # via django
tablib[xlsx]==3.5.0
    # via -r requirements/base.in
text-unidecode==1.3
    # via python-slugify
twilio==8.8.0
    # via -r requirements/base.in
tzdata==2023.3
    # via
    #   celery
    #   django
    #   django-celery-beat
    #   pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   -r requirements/base.in
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.5.0
    # via -r requirements/base.in
wrapt==1.15.0
    # via deprecated
yarl==1.9.2
    # via aiohttp
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
